agents/voice_agent/requirements-voice.txt
fastapi
pydantic
python-dotenv
openai-whisper
gtts
pydub
webrtcvad
diskcache
